requirements.txt
PyAutoGUI==0.9.33
PyHamcrest==1.9.0
pytest==2.9.1